Final Push of MAC_user_friendly
I am still having some issues with the calcDIsp function. It is supposed to extrapolate beam x-component rotation to turn the 13 numerical data points to 26 numerical data points to match the numeber of experimental data points. In a sense it accounts for torsion. It is resulting in near zero rotation, needs further work